chatbot_streamlit/requirements.txt
streamlit>=1.31.0
google-cloud-storage>=2.14.0
google-cloud-aiplatform>=1.42.0
google-generativeai>=0.3.0
pandas>=2.2.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
